Update to Ubuntu 16.04
Seems to work great here for hello_world. Hit some build issues with Serde v0.7.15 on Nightly, but that's not a platform issue ("expected slice, found struct `std::vec::Vec`"). Have to test it on some single board computers here, but the toolchain worked.

Thanks for making this project @alexcrichton. It was much easier to use this than rolling my own was. XD